+++ lib/Streaming/ByteString/Char8.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE OverloadedStrings   #-}
 {-# LANGUAGE RankNTypes          #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE ViewPatterns        #-}
 
 -- |
 -- Module      : Streaming.ByteString.Char8
@@ -26,20 +27,20 @@ module Streaming.ByteString.Char8
   , ByteString
 
     -- * Introducing and eliminating 'ByteStream's
-  , empty            -- empty :: ByteStream m ()
-  , pack             -- pack :: Monad m => String -> ByteStream m ()
+  , empty
+  , pack
   , unpack
   , string
   , unlines
   , unwords
-  , singleton        -- singleton :: Monad m => Char -> ByteStream m ()
-  , fromChunks       -- fromChunks :: Monad m => Stream (Of ByteString) m r -> ByteStream m r
-  , fromLazy         -- fromLazy :: Monad m => ByteString -> ByteStream m ()
-  , fromStrict       -- fromStrict :: ByteString -> ByteStream m ()
-  , toChunks         -- toChunks :: Monad m => ByteStream m r -> Stream (Of ByteString) m r
-  , toLazy           -- toLazy :: Monad m => ByteStream m () -> m ByteString
+  , singleton
+  , fromChunks
+  , fromLazy
+  , fromStrict
+  , toChunks
+  , toLazy
   , toLazy_
-  , toStrict         -- toStrict :: Monad m => ByteStream m () -> m ByteString
+  , toStrict
   , toStrict_
   , effects
   , copy
@@ -47,72 +48,74 @@ module Streaming.ByteString.Char8
   , mwrap
 
     -- * Transforming ByteStreams
-  , map              -- map :: Monad m => (Char -> Char) -> ByteStream m r -> ByteStream m r
-  , intercalate      -- intercalate :: Monad m => ByteStream m () -> Stream (ByteStream m) m r -> ByteStream m r
-  , intersperse      -- intersperse :: Monad m => Char -> ByteStream m r -> ByteStream m r
+  , map
+  , intercalate
+  , intersperse
 
     -- * Basic interface
-  , cons             -- cons :: Monad m => Char -> ByteStream m r -> ByteStream m r
-  , cons'            -- cons' :: Char -> ByteStream m r -> ByteStream m r
+  , cons
+  , cons'
   , snoc
-  , append           -- append :: Monad m => ByteStream m r -> ByteStream m s -> ByteStream m s
-  , filter           -- filter :: (Char -> Bool) -> ByteStream m r -> ByteStream m r
-  , head             -- head :: Monad m => ByteStream m r -> m Char
-  , head_            -- head' :: Monad m => ByteStream m r -> m (Of Char r)
-  , last             -- last :: Monad m => ByteStream m r -> m Char
-  , last_            -- last' :: Monad m => ByteStream m r -> m (Of Char r)
-  , null             -- null :: Monad m => ByteStream m r -> m Bool
+  , append
+  , filter
+  , head
+  , head_
+  , last
+  , last_
+  , null
   , null_
+  , nulls
   , testNull
-  , nulls            -- null' :: Monad m => ByteStream m r -> m (Of Bool r)
-  , uncons           -- uncons :: Monad m => ByteStream m r -> m (Either r (Char, ByteStream m r))
+  , uncons
   , nextChar
   , skipSomeWS
 
     -- * Substrings
     -- ** Breaking strings
-  , break            -- break :: Monad m => (Char -> Bool) -> ByteStream m r -> ByteStream m (ByteStream m r)
-  , drop             -- drop :: Monad m => GHC.Int.Int64 -> ByteStream m r -> ByteStream m r
+  , break
+  , drop
   , dropWhile
-  , group            -- group :: Monad m => ByteStream m r -> Stream (ByteStream m) m r
+  , group
   , groupBy
-  , span             -- span :: Monad m => (Char -> Bool) -> ByteStream m r -> ByteStream m (ByteStream m r)
-  , splitAt          -- splitAt :: Monad m => GHC.Int.Int64 -> ByteStream m r -> ByteStream m (ByteStream m r)
-  , splitWith        -- splitWith :: Monad m => (Char -> Bool) -> ByteStream m r -> Stream (ByteStream m) m r
-  , take             -- take :: Monad m => GHC.Int.Int64 -> ByteStream m r -> ByteStream m ()
-  , takeWhile        -- takeWhile :: (Char -> Bool) -> ByteStream m r -> ByteStream m ()
+  , span
+  , splitAt
+  , splitWith
+  , take
+  , takeWhile
 
     -- ** Breaking into many substrings
-  , split            -- split :: Monad m => Char -> ByteStream m r -> Stream (ByteStream m) m r
+  , split
   , lines
-  , words
   , lineSplit
-  , denull
+  , words
 
     -- ** Special folds
-  , concat          -- concat :: Monad m => Stream (ByteStream m) m r -> ByteStream m r
+  , concat
+  , denull
 
     -- * Builders
   , toStreamingByteString
+
   , toStreamingByteStringWith
+
   , toBuilder
   , concatBuilders
 
     -- * Building ByteStreams
     -- ** Infinite ByteStreams
-  , repeat           -- repeat :: Char -> ByteStream m ()
-  , iterate          -- iterate :: (Char -> Char) -> Char -> ByteStream m ()
-  , cycle            -- cycle :: Monad m => ByteStream m r -> ByteStream m s
+  , repeat
+  , iterate
+  , cycle
 
     -- ** Unfolding ByteStreams
-  , unfoldr          -- unfoldr :: (a -> Maybe (Char, a)) -> a -> ByteStream m ()
-  , unfoldM          -- unfold  :: (a -> Either r (Char, a)) -> a -> ByteStream m r
+  , unfoldr
+  , unfoldM
   , reread
 
     -- *  Folds, including support for `Control.Foldl`
---    , foldr            -- foldr :: Monad m => (Char -> a -> a) -> a -> ByteStream m () -> m a
-  , fold             -- fold :: Monad m => (x -> Char -> x) -> x -> (x -> b) -> ByteStream m () -> m b
-  , fold_            -- fold' :: Monad m => (x -> Char -> x) -> x -> (x -> b) -> ByteStream m r -> m (b, r)
+    -- , foldr
+  , fold
+  , fold_
   , length
   , length_
   , count
@@ -121,29 +124,29 @@ module Streaming.ByteString.Char8
 
     -- * I\/O with 'ByteStream's
     -- ** Standard input and output
-  , getContents      -- getContents :: ByteStream IO ()
-  , stdin            -- stdin :: ByteStream IO ()
-  , stdout           -- stdout :: ByteStream IO r -> IO r
-  , interact         -- interact :: (ByteStream IO () -> ByteStream IO r) -> IO r
+  , getContents
+  , stdin
+  , stdout
+  , interact
   , putStr
   , putStrLn
 
     -- ** Files
-  , readFile         -- readFile :: FilePath -> ByteStream IO ()
-  , writeFile        -- writeFile :: FilePath -> ByteStream IO r -> IO r
-  , appendFile       -- appendFile :: FilePath -> ByteStream IO r -> IO r
+  , readFile
+  , writeFile
+  , appendFile
 
     -- ** I\/O with Handles
-  , fromHandle       -- fromHandle :: Handle -> ByteStream IO ()
-  , toHandle         -- toHandle :: Handle -> ByteStream IO r -> IO r
-  , hGet             -- hGet :: Handle -> Int -> ByteStream IO ()
-  , hGetContents     -- hGetContents :: Handle -> ByteStream IO ()
-  , hGetContentsN    -- hGetContentsN :: Int -> Handle -> ByteStream IO ()
-  , hGetN            -- hGetN :: Int -> Handle -> Int -> ByteStream IO ()
-  , hGetNonBlocking  -- hGetNonBlocking :: Handle -> Int -> ByteStream IO ()
-  , hGetNonBlockingN -- hGetNonBlockingN :: Int -> Handle -> Int -> ByteStream IO ()
-  , hPut             -- hPut :: Handle -> ByteStream IO r -> IO r
---    , hPutNonBlocking  -- hPutNonBlocking :: Handle -> ByteStream IO r -> ByteStream IO r
+  , fromHandle
+  , toHandle
+  , hGet
+  , hGetContents
+  , hGetContentsN
+  , hGetN
+  , hGetNonBlocking
+  , hGetNonBlockingN
+  , hPut
+    -- , hPutNonBlocking
 
     -- * Simple chunkwise operations
   , unconsChunk
@@ -158,10 +161,10 @@ module Streaming.ByteString.Char8
   , chunkMapM_
 
     -- * Etc.
---    , zipWithStream    -- zipWithStream :: Monad m => (forall x. a -> ByteStream m x -> ByteStream m x) -> [a] -> Stream (ByteStream m) m r -> Stream (ByteStream m) m r
-  , distribute      -- distribute :: ByteStream (t m) a -> t (ByteStream m) a
-  , materialize
   , dematerialize
+  , materialize
+  , distribute
+  , zipWithStream
   ) where
 
 import           Prelude hiding
@@ -194,7 +197,7 @@ import           Streaming.ByteString
     length_, nextChunk, null, null_, nulls, readFile, splitAt, stdin, stdout,
     take, testNull, toBuilder, toChunks, toHandle, toLazy, toLazy_,
     toStreamingByteString, toStreamingByteStringWith, toStrict, toStrict_,
-    unconsChunk, writeFile)
+    unconsChunk, writeFile, zipWithStream)
 
 import           Data.Word (Word8)
 import           Foreign.ForeignPtr (withForeignPtr)
@@ -203,7 +206,7 @@ import           Foreign.Storable
 import qualified System.IO as IO
 
 -- | Given a stream of bytes, produce a vanilla `Stream` of characters.
-unpack ::  Monad m => ByteStream m r ->  Stream (Of Char) m r
+unpack :: Monad m => ByteStream m r -> Stream (Of Char) m r
 unpack bs = case bs of
     Empty r    -> Return r
     Go m       -> Effect (fmap unpack m)
@@ -303,17 +306,22 @@ groupBy :: Monad m => (Char -> Char -> Bool) -> ByteStream m r -> Stream (ByteSt
 groupBy rel = Q.groupBy (\w w' -> rel (w2c w) (w2c w'))
 {-# INLINE groupBy #-}
 
--- | /O(1)/ Extract the head and tail of a ByteStream, returning Nothing
--- if it is empty.
+-- | /O(1)/ Extract the head and tail of a 'ByteStream', or its return value if
+-- it is empty. This is the \'natural\' uncons for an effectful byte stream.
 uncons :: Monad m => ByteStream m r -> m (Either r (Char, ByteStream m r))
+uncons (Chunk c@(B.length -> len) cs)
+    | len > 0    = let !h = w2c $ B.unsafeHead c
+                       !t = if len > 1 then Chunk (B.unsafeTail c) cs else cs
+                    in return $ Right (h, t)
+    | otherwise  = uncons cs
+uncons (Go m)    = m >>= uncons
 uncons (Empty r) = return (Left r)
-uncons (Chunk c cs)
-    = return $ Right (w2c (B.unsafeHead c)
-                     , if B.length c == 1
-                         then cs
-                         else Chunk (B.unsafeTail c) cs )
-uncons (Go m) = m >>= uncons
 {-# INLINABLE uncons #-}
+
+nextChar :: Monad m => ByteStream m r -> m (Either r (Char, ByteStream m r))
+nextChar = uncons
+{-# INLINABLE nextChar #-}
+{-# DEPRECATED nextChar "Use uncons instead." #-}
 
 -- ---------------------------------------------------------------------
 -- Transformations
@@ -514,17 +522,13 @@ lines text0 = loop1 text0
 --  the \"lines\" had embedded newlines.
 --
 --  * @'unlines' . 'lines'@ will replace @\\r\\n@ with @\\n@.
-unlines :: Monad m => Stream (ByteStream m) m r ->  ByteStream m r
+unlines :: Monad m => Stream (ByteStream m) m r -> ByteStream m r
 unlines = loop where
   loop str =  case str of
     Return r -> Empty r
     Step bstr   -> do
       st <- bstr
-      let bs = unlines st
-      case bs of
-        Chunk "" (Empty r)   -> Empty r
-        Chunk "\n" (Empty _) -> bs
-        _                    -> cons' '\n' bs
+      cons' '\n' $ unlines st
     Effect m  -> Go (fmap unlines m)
 {-# INLINABLE unlines #-}
 
@@ -662,15 +666,6 @@ count_ c = Q.count_ (c2w c)
 count :: Monad m => Char -> ByteStream m r -> m (Of Int r)
 count c = Q.count (c2w c)
 {-# INLINE count #-}
-
--- | /O(1)/ Extract the head and tail of a 'ByteStream', or its return value if
--- it is empty. This is the \'natural\' uncons for an effectful byte stream.
-nextChar :: Monad m => ByteStream m r -> m (Either r (Char, ByteStream m r))
-nextChar b = do
-  e <- Q.nextByte b
-  case e of
-    Left r       -> return $! Left r
-    Right (w,bs) -> return $! Right (w2c w, bs)
 
 -- | Print a stream of bytes to STDOUT.
 putStr :: MonadIO m => ByteStream m r -> m r